add custom proc cluster bdn
