+++ benchmark/BDN.benchmark/Cluster/RespNoSlot.cs
@@ -1,0 +1,22 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using BenchmarkDotNet.Attributes;
+
+namespace BDN.benchmark.Cluster
+{
+    [MemoryDiagnoser]
+    public unsafe class RespClusterBenchNoSlot : RespClusterBench
+    {
+        public new void GlobalSetup()
+        {
+            enableCluster = false;
+            base.GlobalSetup();
+        }
+
+        public new void GlobalCleanup()
+        {
+            base.GlobalCleanup();
+        }
+    }
+}
